requirements.txt
django
djangorestframework
redis
celery
requests